+++ core/src/test/java/io/cucumber/core/stepexpression/StepExpressionFactoryTest.java
@@ -158,13 +158,23 @@ class StepExpressionFactoryTest {
     }
 
     @Test
-    void docstring_expression_transform_doc_string_with_content_type_to_string() {
+    void docstring_expression_transform_doc_string_to_string() {
         String docString = "A rather long and boring string of documentation";
-        String contentType = "doc";
         StepDefinition stepDefinition = new StubStepDefinition("Given some stuff:", String.class);
         StepExpression expression = stepExpressionFactory.createExpression(stepDefinition);
-        List<Argument> match = expression.match("Given some stuff:", docString, contentType);
+        List<Argument> match = expression.match("Given some stuff:", docString, null);
         assertThat(match.get(0).getValue(), is(equalTo(docString)));
+    }
+
+    @Test
+    void docstring_and_datatable_match_same_step_definition() {
+        String docString = "A rather long and boring string of documentation";
+        StepDefinition stepDefinition = new StubStepDefinition("Given some stuff:", UNKNOWN_TYPE);
+        StepExpression expression = stepExpressionFactory.createExpression(stepDefinition);
+        List<Argument> match = expression.match("Given some stuff:", docString, null);
+        assertThat(match.get(0).getValue(), is(equalTo(DocString.create(docString))));
+        match = expression.match("Given some stuff:", table);
+        assertThat(match.get(0).getValue(), is(equalTo(DataTable.create(table))));
     }
 
     @Test