+++ docstring/src/test/java/io/cucumber/docstring/DocStringTypeRegistryTest.java
@@ -1,22 +1,26 @@
 package io.cucumber.docstring;
 
+import com.fasterxml.jackson.core.TreeNode;
 import com.fasterxml.jackson.databind.JsonNode;
 import org.junit.jupiter.api.Test;
 
 import static org.hamcrest.CoreMatchers.equalTo;
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.collection.IsCollectionWithSize.hasSize;
+import static org.junit.jupiter.api.Assertions.assertAll;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
 class DocStringTypeRegistryTest {
 
+    public static final String DEFAULT_CONTENT_TYPE = "";
     private final DocStringTypeRegistry registry = new DocStringTypeRegistry();
 
     @Test
     void anonymous_doc_string_is_predefined() {
         DocStringType docStringType = new DocStringType(
             String.class,
-            "",
+            DEFAULT_CONTENT_TYPE,
             (String s) -> s);
 
         CucumberDocStringException actualThrown = assertThrows(
@@ -27,10 +31,10 @@ class DocStringTypeRegistryTest {
     }
 
     @Test
-    void doc_string_types_must_be_unique() {
+    void doc_string_types_of_same_content_type_must_have_unique_return_type() {
         registry.defineDocStringType(new DocStringType(
             JsonNode.class,
-            "json",
+            "application/json",
             (String s) -> null));
 
         DocStringType duplicate = new DocStringType(
@@ -42,8 +46,81 @@ class DocStringTypeRegistryTest {
             CucumberDocStringException.class,
             () -> registry.defineDocStringType(duplicate));
         assertThat(exception.getMessage(), is("" +
-                "There is already docstring type registered for 'json' and com.fasterxml.jackson.databind.JsonNode.\n" +
+                "There is already docstring type registered for 'application/json' and com.fasterxml.jackson.databind.JsonNode.\n"
+                +
                 "You are trying to add 'application/json' and com.fasterxml.jackson.databind.JsonNode"));
+    }
+
+    @Test
+    void can_register_multiple_doc_string_with_different_content_type_but_same_return_type() {
+        registry.defineDocStringType(new DocStringType(
+            JsonNode.class,
+            "application/json",
+            (String s) -> null));
+
+        registry.defineDocStringType(new DocStringType(
+            JsonNode.class,
+            "application/xml",
+            (String s) -> null));
+
+        registry.defineDocStringType(new DocStringType(
+            JsonNode.class,
+            "application/yml",
+            (String s) -> null));
+
+        assertThat(registry.lookup(null, JsonNode.class), hasSize(3));
+    }
+
+    @Test
+    void no_content_type_association_is_made() {
+        registry.defineDocStringType(new DocStringType(
+            JsonNode.class,
+            "application/json",
+            (String s) -> null));
+
+        registry.defineDocStringType(new DocStringType(
+            JsonNode.class,
+            "json",
+            (String s) -> null));
+
+        registry.defineDocStringType(new DocStringType(
+            JsonNode.class,
+            "json/application",
+            (String s) -> null));
+
+        assertThat(registry.lookup(null, JsonNode.class), hasSize(3));
+    }
+
+    @Test
+    void can_add_multiple_default_content_types_with_different_return_types() {
+        registry.defineDocStringType(new DocStringType(
+            JsonNode.class,
+            DEFAULT_CONTENT_TYPE,
+            (String s) -> null));
+
+        registry.defineDocStringType(new DocStringType(
+            TreeNode.class,
+            DEFAULT_CONTENT_TYPE,
+            (String s) -> null));
+
+        assertThat(registry.lookup(DEFAULT_CONTENT_TYPE, JsonNode.class), hasSize(1));
+        assertThat(registry.lookup(DEFAULT_CONTENT_TYPE, TreeNode.class), hasSize(1));
+    }
+
+    @Test
+    void can_add_same_content_type_with_different_return_types() {
+        registry.defineDocStringType(new DocStringType(
+            JsonNode.class,
+            "application/json",
+            (String s) -> null));
+
+        registry.defineDocStringType(new DocStringType(
+            TreeNode.class,
+            "application/json",
+            (String s) -> null));
+
+        assertThat(registry.lookup("application/json", JsonNode.class), hasSize(1));
+        assertThat(registry.lookup("application/json", TreeNode.class), hasSize(1));
     }
 
 }