+++ guice/src/main/java/io/cucumber/guice/InjectorSourceFactory.java
@@ -2,26 +2,27 @@ package io.cucumber.guice;
 
 import com.google.inject.Guice;
 import com.google.inject.Stage;
+import io.cucumber.core.logging.Logger;
+import io.cucumber.core.logging.LoggerFactory;
 
 import java.util.Map;
 
-import static java.text.MessageFormat.format;
+import static java.lang.String.format;
 
 final class InjectorSourceFactory {
-
+    private static final Logger log = LoggerFactory.getLogger(GuiceFactory.class);
     static final String GUICE_INJECTOR_SOURCE_KEY = "guice.injector-source";
-    private final Map<String, String> properties;
+    private final Class<?> injectorSourceClass;
 
-    InjectorSourceFactory(Map<String, String> properties) {
-        this.properties = properties;
+    InjectorSourceFactory(Class<?> injectorSourceClass) {
+        this.injectorSourceClass = injectorSourceClass;
     }
 
     InjectorSource create() {
-        String injectorSourceClassName = properties.get(GUICE_INJECTOR_SOURCE_KEY);
-        if (injectorSourceClassName == null) {
+        if (injectorSourceClass == null) {
             return createDefaultScenarioModuleInjectorSource();
         } else {
-            return instantiateUserSpecifiedInjectorSource(injectorSourceClassName);
+            return instantiateUserSpecifiedInjectorSource(injectorSourceClass);
         }
     }
 
@@ -29,13 +30,34 @@ final class InjectorSourceFactory {
         return () -> Guice.createInjector(Stage.PRODUCTION, CucumberModules.createScenarioModule());
     }
 
-    private InjectorSource instantiateUserSpecifiedInjectorSource(String injectorSourceClassName) {
+    private InjectorSource instantiateUserSpecifiedInjectorSource(Class<?> injectorSourceClass) {
         try {
-            return (InjectorSource) Class
-                    .forName(injectorSourceClassName, true, Thread.currentThread().getContextClassLoader())
-                    .newInstance();
+            return (InjectorSource) injectorSourceClass.getConstructor().newInstance();
         } catch (Exception e) {
-            String message = format("Instantiation of ''{0}'' failed. Check the caused by exception and ensure your" +
+            String message = format("Instantiation of '%s' failed. Check the caused by exception and ensure your" +
+                    "InjectorSource implementation is accessible and has a public zero args constructor.",
+                injectorSourceClass.getName());
+            throw new InjectorSourceInstantiationFailed(message, e);
+        }
+    }
+
+    @Deprecated
+    static Class<?> loadInjectorSourceFromProperties(Map<String, String> properties) {
+        String injectorSourceClassName = properties.get(GUICE_INJECTOR_SOURCE_KEY);
+
+        if (injectorSourceClassName == null) {
+            return null;
+        }
+
+        log.warn(
+            () -> format("The '%s' property has been deprecated." +
+                    "Add a class implementing '%s' on the glue path instead",
+                GUICE_INJECTOR_SOURCE_KEY, InjectorSource.class.getName()));
+
+        try {
+            return Class.forName(injectorSourceClassName, true, Thread.currentThread().getContextClassLoader());
+        } catch (Exception e) {
+            String message = format("Instantiation of '%s' failed. Check the caused by exception and ensure your" +
                     "InjectorSource implementation is accessible and has a public zero args constructor.",
                 injectorSourceClassName);
             throw new InjectorSourceInstantiationFailed(message, e);