+++ java/src/test/java/io/cucumber/java/JavaDocStringTypeDefinitionTest.java
@@ -1,5 +1,6 @@
 package io.cucumber.java;
 
+import com.fasterxml.jackson.core.type.TypeReference;
 import io.cucumber.core.backend.Lookup;
 import io.cucumber.docstring.DocString;
 import io.cucumber.docstring.DocStringTypeRegistry;
@@ -7,10 +8,15 @@ import io.cucumber.docstring.DocStringTypeRegistryDocStringConverter;
 import org.junit.jupiter.api.Test;
 
 import java.lang.reflect.Method;
+import java.lang.reflect.Type;
+import java.util.Collections;
+import java.util.Map;
 
 import static org.hamcrest.CoreMatchers.startsWith;
 import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.collection.IsMapContaining.hasEntry;
 import static org.hamcrest.core.Is.is;
+import static org.junit.jupiter.api.Assertions.assertAll;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
 class JavaDocStringTypeDefinitionTest {
@@ -31,13 +37,22 @@ class JavaDocStringTypeDefinitionTest {
     @Test
     void can_define_doc_string_converter() throws NoSuchMethodException {
         Method method = JavaDocStringTypeDefinitionTest.class.getMethod("convert_doc_string_to_string", String.class);
+        JavaDocStringTypeDefinition definition = new JavaDocStringTypeDefinition("text/plain", method, lookup);
+        registry.defineDocStringType(definition.docStringType());
+        assertThat(converter.convert(docString, Object.class), is("some_desired_string"));
+    }
+
+    @Test
+    void can_define_doc_string_without_content_types_converter() throws NoSuchMethodException {
+        Method method = JavaDocStringTypeDefinitionTest.class.getMethod("convert_doc_string_to_string", String.class);
         JavaDocStringTypeDefinition definition = new JavaDocStringTypeDefinition("", method, lookup);
         registry.defineDocStringType(definition.docStringType());
-        assertThat(converter.convert(docString, Object.class), is("convert_doc_string_to_string"));
+        assertThat(converter.convert(DocString.create("some doc string"), Object.class),
+            is("some_desired_string"));
     }
 
     public Object convert_doc_string_to_string(String docString) {
-        return "convert_doc_string_to_string";
+        return "some_desired_string";
     }
 
     @Test
@@ -75,11 +90,52 @@ class JavaDocStringTypeDefinitionTest {
 
     @Test
     void must_return_something() throws NoSuchMethodException {
-        Method method = JavaDocStringTypeDefinitionTest.class.getMethod("converts_string_to_void", String.class);
-        assertThrows(InvalidMethodSignatureException.class, () -> new JavaDocStringTypeDefinition("", method, lookup));
+        Method voidMethod = JavaDocStringTypeDefinitionTest.class.getMethod("converts_string_to_void", String.class);
+        Method voidObjectMethod = JavaDocStringTypeDefinitionTest.class.getMethod("converts_string_to_void_object",
+            String.class);
+
+        assertAll(
+            () -> assertThrows(InvalidMethodSignatureException.class,
+                () -> new JavaDocStringTypeDefinition("", voidMethod, lookup)),
+            () -> assertThrows(InvalidMethodSignatureException.class,
+                () -> new JavaDocStringTypeDefinition("", voidObjectMethod, lookup)));
     }
 
     public void converts_string_to_void(String docString) {
+    }
+
+    public Void converts_string_to_void_object(String docString) {
+        return null;
+    }
+
+    @Test
+    public void correct_conversion_is_used_for_simple_and_complex_return_types() throws NoSuchMethodException {
+        Method simpleMethod = JavaDocStringTypeDefinitionTest.class.getMethod("converts_string_to_simple_type",
+            String.class);
+        JavaDocStringTypeDefinition simpleDefinition = new JavaDocStringTypeDefinition("text/plain", simpleMethod,
+            lookup);
+        registry.defineDocStringType(simpleDefinition.docStringType());
+
+        Method complexMethod = JavaDocStringTypeDefinitionTest.class.getMethod("converts_string_to_complex_type",
+            String.class);
+        JavaDocStringTypeDefinition complexDefinition = new JavaDocStringTypeDefinition("text/plain", complexMethod,
+            lookup);
+        registry.defineDocStringType(complexDefinition.docStringType());
+
+        Type simpleType = Map.class;
+        assertThat(converter.convert(docString, simpleType), hasEntry("some_simple_type", Collections.emptyMap()));
+        Type complexType = new TypeReference<Map<String, Map<String, String>>>() {
+        }.getType();
+        assertThat(converter.convert(docString, complexType), hasEntry("some_complex_type", Collections.emptyMap()));
+    }
+
+    @SuppressWarnings("rawtypes")
+    public Map converts_string_to_simple_type(String docString) {
+        return Collections.singletonMap("some_simple_type", Collections.emptyMap());
+    }
+
+    public Map<String, Map<String, String>> converts_string_to_complex_type(String docString) {
+        return Collections.singletonMap("some_complex_type", Collections.emptyMap());
     }
 
 }