+++ java8/src/main/java/io/cucumber/java8/LambdaGlue.java
@@ -406,7 +406,7 @@ public interface LambdaGlue {
      *                    <code>type</code> from the doc string
      * @see               io.cucumber.docstring.DocStringType
      */
-    default void DocStringType(String contentType, DocStringDefinitionBody<?> body) {
+    default <T> void DocStringType(String contentType, DocStringDefinitionBody<T> body) {
         LambdaGlueRegistry.INSTANCE.get().addDocStringType(new Java8DocStringTypeDefinition(contentType, body));
     }